+++ darts/models/forecasting/sklearn_model.py
@@ -67,6 +67,7 @@ from darts.utils.historical_forecasts import (
 from darts.utils.likelihood_models.sklearn import QuantileRegression, SKLearnLikelihood
 from darts.utils.multioutput import MultiOutputRegressor
 from darts.utils.ts_utils import get_single_series, seq2series, series2seq
+from darts.utils.utils import random_method
 
 logger = get_logger(__name__)
 
@@ -77,6 +78,7 @@ FUTURE_LAGS_TYPE = Union[
 
 
 class SKLearnModel(GlobalForecastingModel):
+    @random_method
     def __init__(
         self,
         lags: Optional[LAGS_TYPE] = None,
@@ -88,6 +90,7 @@ class SKLearnModel(GlobalForecastingModel):
         model=None,
         multi_models: Optional[bool] = True,
         use_static_covariates: bool = True,
+        random_state: Optional[int] = None,
     ):
         """Regression Model
         Can be used to fit any scikit-learn-like regressor class to predict the target time series from lagged values.
@@ -178,6 +181,8 @@ class SKLearnModel(GlobalForecastingModel):
             Whether the model should use static covariate information in case the input `series` passed to ``fit()``
             contain static covariates. If ``True``, and static covariates are available at fitting time, will enforce
             that all target `series` have the same static covariate dimensionality in ``fit()`` and ``predict()``.
+        random_state
+            Controls the model randomness for reproducible forecasting.
 
         Examples
         --------
@@ -911,7 +916,7 @@ class SKLearnModel(GlobalForecastingModel):
                 "eval_weight_name": val_weight_name,
                 "n_jobs": n_jobs_multioutput_wrapper,
             }
-            self.model = MultiOutputRegressor(self.model, **mor_kwargs)
+            self.model = MultiOutputRegressor(estimator=self.model, **mor_kwargs)
 
         if (
             not isinstance(self.model, MultiOutputRegressor)
@@ -998,6 +1003,7 @@ class SKLearnModel(GlobalForecastingModel):
         verbose: bool = False,
         predict_likelihood_parameters: bool = False,
         show_warnings: bool = True,
+        random_state: Optional[int] = None,
         **kwargs,
     ) -> Union[TimeSeries, Sequence[TimeSeries]]:
         """Forecasts values for `n` time steps after the end of the series.
@@ -1027,6 +1033,8 @@ class SKLearnModel(GlobalForecastingModel):
             Default: ``False``
         show_warnings
             Optionally, control whether warnings are shown. Not effective for all models.
+        random_state
+            Controls the randomness of probabilistic predictions.
         **kwargs : dict, optional
             Additional keyword arguments passed to the `predict` method of the model. Only works with
             univariate target series.
@@ -1216,7 +1224,11 @@ class SKLearnModel(GlobalForecastingModel):
 
             # X has shape (n_series * n_samples, n_regression_features)
             prediction = self._predict(
-                X, num_samples, predict_likelihood_parameters, **kwargs
+                X,
+                num_samples,
+                predict_likelihood_parameters,
+                random_state=random_state,
+                **kwargs,
             )
             # prediction shape (n_series * n_samples, output_chunk_length, n_components)
             # append prediction to final predictions
@@ -1250,11 +1262,13 @@ class SKLearnModel(GlobalForecastingModel):
 
         return predictions[0] if called_with_single_series else predictions
 
+    @random_method
     def _predict(
         self,
         x: np.ndarray,
         num_samples: int,
         predict_likelihood_parameters: bool,
+        random_state: Optional[int] = None,
         **kwargs,
     ) -> np.ndarray:
         """Generate predictions.
@@ -1473,6 +1487,7 @@ class SKLearnModelWithCategoricalCovariates(SKLearnModel, ABC):
         categorical_past_covariates: Optional[Union[str, list[str]]] = None,
         categorical_future_covariates: Optional[Union[str, list[str]]] = None,
         categorical_static_covariates: Optional[Union[str, list[str]]] = None,
+        random_state: Optional[int] = None,
     ):
         """
         Extension of `SKLearnModel` for regression models that support categorical covariates.
@@ -1570,6 +1585,8 @@ class SKLearnModelWithCategoricalCovariates(SKLearnModel, ABC):
         categorical_static_covariates
             Optionally, string or list of strings specifying the static covariates that should be treated as
             categorical.
+        random_state
+            Controls the model randomness for reproducible forecasting.
         """
         super().__init__(
             lags=lags,
@@ -1581,6 +1598,7 @@ class SKLearnModelWithCategoricalCovariates(SKLearnModel, ABC):
             model=model,
             multi_models=multi_models,
             use_static_covariates=use_static_covariates,
+            random_state=random_state,
         )
 
         if categorical_static_covariates is not None and not use_static_covariates:
@@ -1798,6 +1816,7 @@ class RegressionModel(SKLearnModel):
         model=None,
         multi_models: Optional[bool] = True,
         use_static_covariates: bool = True,
+        random_state: Optional[int] = None,
     ):
         """Regression Model
         Can be used to fit any scikit-learn-like regressor class to predict the target time series from lagged values.
@@ -1891,6 +1910,8 @@ class RegressionModel(SKLearnModel):
             Whether the model should use static covariate information in case the input `series` passed to ``fit()``
             contain static covariates. If ``True``, and static covariates are available at fitting time, will enforce
             that all target `series` have the same static covariate dimensionality in ``fit()`` and ``predict()``.
+        random_state
+            Controls the model randomness for reproducible forecasting.
 
         Examples
         --------
@@ -1938,4 +1959,5 @@ class RegressionModel(SKLearnModel):
             model=model,
             multi_models=multi_models,
             use_static_covariates=use_static_covariates,
+            random_state=random_state,
         )