+++ darts/models/forecasting/theta.py
@@ -158,6 +158,7 @@ class Theta(LocalForecastingModel):
         num_samples: int = 1,
         verbose: bool = False,
         show_warnings: bool = True,
+        random_state: Optional[int] = None,
     ) -> "TimeSeries":
         super().predict(n, num_samples)
 
@@ -405,6 +406,7 @@ class FourTheta(LocalForecastingModel):
         num_samples: int = 1,
         verbose: bool = False,
         show_warnings: bool = True,
+        random_state: Optional[int] = None,
     ) -> "TimeSeries":
         super().predict(n, num_samples)