+++ darts/tests/models/forecasting/test_sklearn_models.py
@@ -1435,6 +1435,27 @@ class TestSKLearnModels:
         else:
             assert not isinstance(model.model, MultiOutputRegressor)
 
+    def test_model_representation(self):
+        """Check that model representation works with and without MultiOutputRegressor"""
+        model_1 = LinearRegressionModel(lags=4, output_chunk_length=1)
+        model_1.fit(series=self.sine_univariate1)
+        assert not isinstance(model_1.model, MultiOutputRegressor)
+
+        model_2 = XGBModel(
+            lags=4,
+            output_chunk_length=2,
+            multi_models=True,
+            likelihood="quantile",
+            quantiles=[0.1, 0.5, 0.9],
+            **xgb_test_params,
+        )
+        model_2.fit(series=self.sine_univariate1)
+        assert isinstance(model_2.model, MultiOutputRegressor)
+
+        for model in [model_1, model_2]:
+            assert model.__repr__().startswith(model.__class__.__name__)
+            assert model.__str__().startswith(model.model.__class__.__name__)
+
     def test_get_estimator_multi_models(self):
         """Craft training data so that estimator_[i].predict(X) == i + 1"""
 
@@ -3973,7 +3994,6 @@ class TestProbabilisticSKLearnModels:
             _ = _get_likelihood(
                 likelihood="does_not_exist",
                 n_outputs=1,
-                random_state=None,
                 quantiles=None,
             )
         assert (
@@ -4007,6 +4027,29 @@ class TestProbabilisticSKLearnModels:
         # test whether the next prediction of the same model is different
         pred3 = model.predict(n=10, num_samples=2).values()
         assert (pred2 != pred3).any()
+
+        # test whether two consecutive predictions with a random_state specified are the same
+        pred4 = model.predict(n=10, num_samples=2, random_state=38).values()
+        pred5 = model.predict(n=10, num_samples=2, random_state=38).values()
+        assert (pred4 == pred5).all()
+
+        # additional tests :
+        model = model_cls(**model_kwargs)
+        model.fit(self.constant_noisy_multivar_ts)
+        pred6 = model.predict(n=10, num_samples=2).values()
+        pred7 = model.predict(n=10, num_samples=2).values()
+
+        model = model_cls(**model_kwargs)
+        model.fit(self.constant_noisy_multivar_ts)
+        pred8 = model.predict(n=10, num_samples=2).values()
+        pred9 = model.predict(n=10, num_samples=2, random_state=38).values()
+        pred10 = model.predict(n=10, num_samples=2, random_state=38).values()
+        pred11 = model.predict(n=10, num_samples=2).values()
+
+        assert (pred6 != pred7).any()
+        assert (pred8 == pred6).all()
+        assert (pred9 == pred10).all()
+        assert (pred11 == pred7).all()
 
     @pytest.mark.parametrize("config", product(models_cls_kwargs_errs, [True, False]))
     def test_probabilistic_forecast_accuracy_univariate(self, config):